refactored schema

+++ db/migrate/20190918201515_create_digs.rb
@@ -4,8 +4,8 @@ class CreateDigs < ActiveRecord::Migration[5.2]
       t.belongs_to :kit, foreign_key: true
       t.belongs_to :user, foreign_key: true
       t.string :name
-      t.integer :dig_pad_start
-      t.integer :dig_pad_end
+      t.float :dig_pad_start
+      t.float :dig_pad_end
 
       t.timestamps
     end
